+++ test/array.js
@@ -295,4 +295,30 @@ describe('Array', function() {
 		});
 	});
 
+	describe('#createIterator()', function() {
+
+		var arr = ['a', 'b', 'c', 'd'];
+
+		it('should return an iterator', function() {
+
+			var iter = arr.createIterator();
+
+			assert.equal('Iterator', iter.constructor.name);
+		});
+
+		it('should iterate', function() {
+
+			var iter = arr.createIterator(),
+			    val,
+			    abc = '';
+
+			while (iter.hasNext()) {
+				val = iter.next().value;
+				abc += val;
+			}
+
+			assert.equal('abcd', abc);
+		});
+	});
+
 });